// frontend/history-ui/src/lib/types/DocumentVersionWithoutContent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type DocumentVersionWithoutContent = {
    vaultUpdateId: number;
    documentId: string;
    relativePath: string;
    updatedDate: string;
    isDeleted: boolean;
    userId: string;
    deviceId: string;
    contentSize: number;
};